requirements.txt
Django

Pillow==2.2.1
django-extensions==1.2.2

django-generic-m2m==0.2.1
django-model-utils==1.5.0
django-polymorphic==0.5.3

jsonfield==0.9.19
django-jsonfield==0.9.10

humanize==0.5
factory-boy==2.1.2

psycopg2==2.5.1

python-dateutil==2.1
psutil==1.1.3
marshmallow==0.3.0
shortuuid==0.4
MySQL-python==1.2.4
pysqlite==2.6.3
